webnews/requirement.txt
Django==2.2
djangorestframework==3.9.2
mysql-connector==2.2.9
mysqlclient==1.4.2.post1
Pillow==6.0.0
pkg-resources==0.0.0
pytz==2018.9
sqlparse==0.3.0
